+++ R/zzz.R
@@ -152,7 +152,7 @@ NULL
 NULL
 
 .onAttach <- function(libname, pkgname) {
-  vers <-  "v1.16.0"
+  vers <-  "v1.17.0"
   packageStartupMessage("|=========================================================|")
   packageStartupMessage("| Multi-Environment Trial Analysis (metan) ", vers, "        |")
   packageStartupMessage("| Author: Tiago Olivoto                                   |")